requirements.txt
opencv-python==4.6.0.66
tensorflow==2.9.1
tensorboard==2.9.1
flask ==2.2.2
flask_cors==3.0.10
numpy == 1.23.1
pandas ==1.4.3
Pillow==9.2.0